🚚 Rename useScreenSwitchState to useScreenSwitchStore
